+++ express-zod-api/src/metadata.ts
@@ -1,4 +1,3 @@
-import { globalRegistry } from "zod";
 import type { $ZodType } from "zod/v4/core";
 
 export const metaSymbol = Symbol.for("express-zod-api");
@@ -12,22 +11,4 @@ export const getBrand = (subject: $ZodType) => {
   )
     return brand;
   return undefined;
-};
-
-/**
- * @since zod 3.25.44 can be an object
- * @link https://github.com/colinhacks/zod/pull/4586
- * @since zod 3.25.68 and 4.0.0 was completely removed
- * @link https://github.com/colinhacks/zod/commit/ee5615d76b93aac15d7428a17b834a062235f6a1
- * */
-export const getExamples = (subject: $ZodType): ReadonlyArray<unknown> => {
-  const { examples, example } = globalRegistry.get(subject) || {};
-  if (examples) {
-    return Array.isArray(examples)
-      ? examples
-      : /** @todo remove this branch in v25 */
-        Object.values(examples).map(({ value }) => value);
-  }
-  /** @todo remove this in v25 */
-  return example === undefined ? [] : [example];
 };