+++ express-zod-api/src/result-handler.ts
@@ -9,7 +9,6 @@ import { FlatObject, isObject } from "./common-helpers";
 import { contentTypes } from "./content-type";
 import { IOSchema } from "./io-schema";
 import { ActualLogger } from "./logger-helpers";
-import { getExamples } from "./metadata";
 import {
   DiscriminatedResult,
   ensureHttpError,
@@ -101,8 +100,8 @@ export const defaultResultHandler = new ResultHandler({
       status: z.literal("success"),
       data: output,
     });
-    const examples = getExamples(output); // pulling down:
-    if (examples.length) {
+    const { examples } = globalRegistry.get(output) || {}; // pulling down:
+    if (examples?.length) {
       globalRegistry.add(responseSchema, {
         examples: examples.map((data) => ({
           status: "success" as const,
@@ -152,16 +151,18 @@ export const arrayResultHandler = new ResultHandler({
       output.shape.items instanceof z.ZodArray
         ? output.shape.items
         : z.array(z.any());
-    if (getExamples(responseSchema).length) return responseSchema; // has examples on the items, or pull down:
-    const examples = getExamples(output)
-      .filter(
+    if (globalRegistry.get(responseSchema)?.examples?.length)
+      return responseSchema; // has examples on the items, or pull down:
+    const examples = globalRegistry
+      .get(output)
+      ?.examples?.filter(
         (example): example is { items: unknown[] } =>
           isObject(example) &&
           "items" in example &&
           Array.isArray(example.items),
       )
       .map((example) => example.items);
-    if (examples.length) {
+    if (examples?.length) {
       const current = responseSchema.meta();
       globalRegistry
         .remove(responseSchema) // reassign to avoid cloning